fix minor errors

+++ src/dataload/contrib/snpeff/__init__.py
@@ -33,10 +33,10 @@ def get_mapping():
                             "analyzer": "string_lowercase"
                         },
                         "rank": {
-                            "type": "long"
+                            "type": "integer"
                         },
                         "total": {
-                            "type": "long"
+                            "type": "integer"
                         },
                         "hgvs.c": {
                             "type": "string",
@@ -49,35 +49,35 @@ def get_mapping():
                         "cdna": {
                             "properties": {
                                 "position": {
-                                    "type": "long"
+                                    "type": "integer"
                                 },
                                 "length": {
-                                    "type": "long"
+                                    "type": "integer"
                                 }
                             }
                         },
                         "cds": {
                             "properties": {
                                 "position": {
-                                    "type": "long"
+                                    "type": "integer"
                                 },
                                 "length": {
-                                    "type": "long"
+                                    "type": "integer"
                                 }
                             }
                         },
                         "protein": {
                             "properties": {
                                 "position": {
-                                    "type": "long"
+                                    "type": "integer"
                                 },
                                 "length": {
-                                    "type": "long"
+                                    "type": "integer"
                                 }
                             }
                         },
                         "distance_to_feature": {
-                            "type": "long"
+                            "type": "integer"
                         }
                     }
                 },
